@@ -8,7 +8,7 @@ namespace TimeTracker
     public static class Extensions
     {
         public static string AppData(this ITauronEnviroment enviroment)
-            => Path.Combine(enviroment.LocalApplicationData, "Time-Tracker");
+            => Path.Combine(enviroment.LocalApplicationData.OriginalPath, "Time-Tracker");
 
         public static void ReportError(this IEventAggregator aggregator, Exception exception)
             => aggregator.GetEvent<ErrorCarrier, Exception>().Publish(exception);
