@@ -11,7 +11,7 @@ namespace Tauron.Application.Wpf.UI;
 [PublicAPI]
 public sealed class Loc : UpdatableMarkupExtension
 {
-    private static readonly Dictionary<string, object?> Cache = new();
+    private static readonly Dictionary<string, object?> Cache = new(StringComparer.Ordinal);
 
     public Loc(string entryName) => EntryName = entryName;
 
